Increases BulletproofGens party_size to allow large transactions

+++ transaction/core/src/validation/validate.rs
@@ -57,7 +57,7 @@ pub fn validate<R: RngCore + CryptoRng>(
 /// The transaction must have at least one input, and no more than the maximum allowed number of inputs.
 fn validate_number_of_inputs(
     tx_prefix: &TxPrefix,
-    maximum_allowed_inputs: u16,
+    maximum_allowed_inputs: u64,
 ) -> TransactionValidationResult<()> {
     let num_inputs = tx_prefix.inputs.len();
 
@@ -77,7 +77,7 @@ fn validate_number_of_inputs(
 /// The transaction must have at least one output.
 fn validate_number_of_outputs(
     tx_prefix: &TxPrefix,
-    maximum_allowed_outputs: u16,
+    maximum_allowed_outputs: u64,
 ) -> TransactionValidationResult<()> {
     let num_outputs = tx_prefix.outputs.len();
